Add `vectorize` flag to torch.autograd.functional.{jacobian, hessian}
Fixes #50584
Add a vectorize flag to torch.autograd.functional.jacobian and
torch.autograd.functional.hessian (default: False). Under the hood, the
vectorize flag uses vmap as the backend to compute the jacobian and
hessian, respectively, providing speedups to users.

The details of this can be a little complicated. "NOTE: [Computing
jacobian with vmap and grad for multiple outputs]" in the code explains
what is going on.

Test Plan:
- I updated all of the jacobian and hessian tests to also use
vectorized=True
- Unit testing for the `_test_construct_standard_basis_for` helper
function
- I added some simple sanity check tests that check e.g. jacobian with
vectorized=False vs
jacobian with vectorized=True. The sanity tests include:
  - functions with multiple inputs
  - functions with outputs(s) that are unrelated to their inputs
  - functions with multiple outputs (jacobian only)
  - functions with multiple inputs and multiple outputs (jacobian only)
  - inputs or outputs that are zero-dim (jacobian only)
  - outputs that are on different devices (jacobian only)
  - outputs that have different dtypes (jacobian only)
- The mechanism for vectorized=True goes through batched gradient
computation. We have separate tests for those (see other PRs in this
stack).

ghstack-source-id: 6ff412d131c8087b3bbcf8fa34133950e16c750c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50915